refactoring: guest-app의 Hook을 hooks 폴더로 이동
