+++ frontend/guest-app/src/hooks/useStringState.js
